Returns the config server dependency (#3)
TL;DR
-----

Adds back the config server dependency so the code will compile

Details
-------

When I added in the Kubernetes config provider I assumed I could
also remove the config server provider. That turned out not to be
the case since some startup code referenced classes for the config
server. Feels quicker to return the dependency than to start
updating the code that references it.