Updates condition for dropping/creating DB (#6)
TL;DR
-----

Updates the check for whether to drop/created the database to
check an environment variable `INIT`

Details
-------

Uses and environment variable to determine whether to initialize
the database. When the variables (`INIT`) is set, then drop the
database and re-created it. Otherwise, assume it is in place and
healthy. This allows for setting the variable `INIT` in a job
used to initialize the service. This pattern is used by all four
service.

Each service also has a check in the main method for the same
variable so that it will run the database seeding code and exit
when initializating.