Simplifies deployment artifacts (#8)
TL;DR
-----

Gets rid of original deployment artifacts and adds our k8s config

Details
-------

To simplify the use of the config server, this change moves the
Kubernets config artifacts originally in the conmfiguration repo
into the main source repo. To simplify the repo for its core
purpose, which is running on Kubernetes, also removes all the older
deployment artifacts originally provided by the base sample repo
this one is based on.