Finds config by specifying `main` branch (#9)
TL;DR
-----

Directly specifies the `main` branch to find the config on the
config server

Details
-------

After wrestling with it for a day or so, I realized that the
reason I wasn't getting any configs was that I've adopted the
modern convention of naming the default, primary branch in my
repositories `main` instead the troublesome `master`. Since
Spring Cloud Config has not caught up with 2020's reckoning with
terms like "master", it would not find the config unless I
explicitly stated that the config client wants its configuration
from the branch named main. This changes the client configuration
to do that.