Corrects the datatype for ingress port (#13)
TL;DR
-----

Uses a number rather than a string to specify the ingress port

Details
-------

Fixes an error on the Ingress controller that was using a string
instead of an integer to specify the port. This was an artifact of
an old template I was using that I incompletely upgraded from the
beta Ingress API to the final version.