Specifies the correct cluster issuer (#14)
TL;DR
-----

Fixes the Ingress object to use the correct cluster issuer

Details
-------

Replaces the cluster issuer annotation on the Music Store Ingress
to use the issue that is configured within the cluster vs. the
default in the template I used to generate the Ingress object. The
name in the template was the one I was using in my home lab which
wasn't right.