Adds variables to manifest
TL;DR
-----

Uses additional variables in the K8s manifest to allow more
explict references in CI/CD

Details
-------

To support a more robust CD process I needed to be able tp specify
a specific image digest for each service and vary the host in the
Ingress object. Uses new YTT data values to allow each deployment/
job to use a specified image and to enable the hostname on the
ingress to be set at deployment time.