Uses variables for images name/tag
TL;DR
-----

Allow for varying the image and tag for each service

Details
-------

To support a more robust CD process I needed to be able tp specify
a specific image digest for each service. This change updates the
Kubernetes manifest to allow both the image name and tag to vary
for each service image.