Switches to a single parameter for images
TL;DR
-----

Uses a single parameter for each image rather than registry,
project, and image components

Details
-------

Reverses course on how to specify the images for each deployment
and job. I was using three component parameters for the registry,
project, and image but ran into difficulty with then when I
switched the pipeline to read the image from the build service
Kubernetes objects (since the status has the entire reference as
one field).

With this change, the value of the image from the `Image` object
status can be used directly without breaking it into pieces.